requirements.txt
moderngl==5.6.4
moderngl_window==2.4.2
Pillow==9.4.0
PyQt6==6.4.0
pyrr==0.10.3
